html/semantics/forms/the-input-element/email.html

<!DOCTYPE html>
<title>Input Email</title>
<link rel="author" title="Kazuki Kanamori" href="mailto:yogurito@gmail.com">
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/master/#e-mail-state-(type=email)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<input type="email" id="single_email" value="user@example.com"/>
<input type="email" id="multiple_email" value="user1@example.com, user2@test.com" multiple/>
<div id="log"></div>

<script type="text/javascript">
  var single = document.getElementById('single_email'),
      mult = document.getElementById('multiple_email'),
      validEmailRegexp = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/;  // from spec

  test(function(){
    assert_false(single.multiple);
  }, "single_email doesn't have the multiple attribute");

  test(function(){
    single.value = '';
    assert_equals(single.value, '');
  }, 'User agents should allow the user to set the value to the empty string');

  test(function(){
    single.value = 'user2@example.com\u000A';
    assert_equals(single.value, 'user2@example.com');
    single.value = 'user3@example.com\u000D';
    assert_equals(single.value, 'user3@example.com');
  }, 'User agents must not allow users to insert "LF" (U+000A) and "CR" (U+000D) characters into the value');

  test(function(){
    single.value = 'user4@example.com';
    assert_true(validEmailRegexp.test(single.value));
    assert_true(single.validity.valid);
    single.value = 'example.com';
    assert_false(validEmailRegexp.test(single.value));
    assert_false(single.validity.valid);
  }, 'The value attribute, if specified and not empty, must have a value that is a single valid e-mail address');

  test(function(){
    single.setAttribute('multiple', true);
    single.value = '  user@example.com  , user2@example.com  ';
    assert_equals(single.value, 'user@example.com,user2@example.com');
    single.removeAttribute('multiple');
    assert_equals(single.value, 'user@example.com,user2@example.com');
  }, 'When the multiple attribute is removed, the user agent must run the value sanitization algorithm');

  test(function(){
    assert_true(mult.multiple);
  }, "multiple_email has the multiple attribute");

  test(function(){
    var values = mult.value;
    assert_equals(mult.value, "user1@example.com,user2@test.com");
  }, "The element's value is the result of splitting on commas the element's value.");

  test(function(){
    mult.value = '  user1@example.com  , user2@test.com, user3@test.com  ';
    assert_equals(mult.value, 'user1@example.com,user2@test.com,user3@test.com');
  }, "run the value sanitization algorithm after setting a new value");

  test(function(){
    mult.value = 'user1@example.com,user2@test.com,user3@test.com';
    var values = mult.value.split(',');
    values.forEach(function(e) { 
      assert_true(validEmailRegexp.test(e));
    });
    assert_true(mult.validity.valid);

    mult.value = 'u,ser1@example.com,user2@test.com,user3@test.com';
    assert_false(mult.validity.valid);
    var values = mult.value.split(',');
    assert_false(validEmailRegexp.test(values[0]));
  }, "valid value is a set of valid email addresses separated by a single ','");
</script>
